+++ sdk/eventhub/event-hubs/src/eventHubConsumerClientModels.ts
@@ -3,6 +3,7 @@ import { ReceivedEventData } from "./eventData";
 import { LastEnqueuedEventProperties } from "./eventHubReceiver";
 import { EventPosition } from "./eventPosition";
 import { TracingOptions } from "./util/operationOptions";
+import { MessagingError } from "@azure/core-amqp";
 
 /**
  * @internal
@@ -29,7 +30,9 @@ export interface BasicPartitionProperties {
 }
 
 /**
- * Provides a set of basic information about the partition as well as the
+ * Interface that describes the context passed to each of the functions that are a part
+ * of the `SubscriptionEventHandlers`. When implementing any of these functions, use
+ * the context object to get information about the partition as well as the
  * ability to checkpoint.
  */
 export interface PartitionContext {
@@ -70,8 +73,7 @@ export interface PartitionContext {
 }
 
 /**
- * Event handler called when events are received. The `context` parameter can be
- * used to get partition information as well as to checkpoint.
+ * Signature of the user provided function invoked by `EventHubConsumerClient` when a set of events is received.
  */
 export type ProcessEventsHandler = (
   events: ReceivedEventData[],
@@ -79,61 +81,59 @@ export type ProcessEventsHandler = (
 ) => Promise<void>;
 
 /**
- * Called when errors occur during event receiving.
+ * Signature of the user provided function invoked by `EventHubConsumerClient` for errors that occur when
+ * receiving events or when executing any of the user provided functions passed to the `subscribe()` method.
  */
-export type ProcessErrorHandler = (error: Error, context: PartitionContext) => Promise<void>;
+export type ProcessErrorHandler = (
+  error: Error | MessagingError,
+  context: PartitionContext
+) => Promise<void>;
 
 /**
- * Called when we first start processing events from a partition.
+ * Signature of the user provided function invoked by `EventHubConsumerClient` just before starting to receive 
+ * events from a partition.
  */
 export type ProcessInitializeHandler = (context: PartitionContext) => Promise<void>;
 
 /**
- * Called when we stop processing events from a partition.
+ * Signature of the user provided function invoked by `EventHubConsumerClient` just after stopping to receive
+ * events from a partition.
  */
 export type ProcessCloseHandler = (reason: CloseReason, context: PartitionContext) => Promise<void>;
 
 /**
- * Optional event handlers that provide more context when subscribing to events.
+ * Interface that describes the functions to be implemented by the user which are invoked by
+ * the `EventHubConsumerClient` when the `subscribe()` method is called to receive events
+ * from Event Hub.
  */
 export interface SubscriptionEventHandlers {
   /**
-   * Event handler called when events are received.
+   * The function invoked by `EventHubConsumerClient` when a set of events is received.
    */
   processEvents: ProcessEventsHandler;
   /**
-   * Called when errors occur during event receiving.
+   * The function invoked by `EventHubConsumerClient` for errors that occur when receiving events
+   * or when executing any of the user provided functions passed to the `subscribe()` method.
    */
   processError: ProcessErrorHandler;
   /**
-   * Called when we first start processing events from a partition.
+   * The function invoked by `EventHubConsumerClient` just before starting to receive events from
+   * a partition. Use this to carry out any setup tasks you would like to have before the client
+   * starts processing a partition.
    */
   processInitialize?: ProcessInitializeHandler;
   /**
-   * Called when we stop processing events from a partition.
+   * The function invoked by `EventHubConsumerClient` just before stopping to receive events from
+   * a partition. Use this to carry out any teardown tasks you would like to have for the partition.
    */
   processClose?: ProcessCloseHandler;
 }
 
 /**
- * Options for subscribe.
+ * Options to configure the `subscribe` method on the `EventHubConsumerClient`.
+ * For example, `{ maxBatchSize: 20, maxWaitTimeInSeconds: 120, startPosition: { sequenceNumber: 123 } }
  */
 export interface SubscribeOptions extends TracingOptions {
-  /**
-   * @property
-   * Indicates whether or not the consumer should request information on the last enqueued event on its
-   * associated partition, and track that information as events are received.
-
-   * When information about the partition's last enqueued event is being tracked, each event received 
-   * from the Event Hubs service will carry metadata about the partition that it otherwise would not. This results in a small amount of
-   * additional network bandwidth consumption that is generally a favorable trade-off when considered
-   * against periodically making requests for partition properties using the Event Hub client.
-   */
-  trackLastEnqueuedEventProperties?: boolean;
-  /**
-   * The owner level to use as this subscription subscribes to partitions.
-   */
-  ownerLevel?: number;
   /**
    * The number of events to request per batch
    */
@@ -149,10 +149,25 @@ export interface SubscribeOptions extends TracingOptions {
    * position for each partition.
    */
   startPosition?: EventPosition | { [partitionId: string]: EventPosition };
+  /**
+   * @property
+   * Indicates whether or not the consumer should request information on the last enqueued event on its
+   * associated partition, and track that information as events are received.
+
+   * When information about the partition's last enqueued event is being tracked, each event received 
+   * from the Event Hubs service will carry metadata about the partition that it otherwise would not. This results in a small amount of
+   * additional network bandwidth consumption that is generally a favorable trade-off when considered
+   * against periodically making requests for partition properties using the Event Hub client.
+   */
+  trackLastEnqueuedEventProperties?: boolean;
+  /**
+   * The owner level to use as this subscription subscribes to partitions.
+   */
+  ownerLevel?: number;
 }
 
 /**
- * Represents the status of a subscribe() call and can be used to stop a subscription.
+ * Interface that describes the object returned by the `subscribe()` method on the `EventHubConsumerClient`.
  */
 export interface Subscription {
   /**