+++ sdk/keyvault/keyvault-certificates/samples/typescript/src/helloWorld.ts
@@ -1,10 +1,17 @@
-import { CertificateClient, DefaultCertificatePolicy } from "../../src";
+// Copyright (c) Microsoft corporation.
+// Licensed under the MIT license.
+
+import { CertificateClient, DefaultCertificatePolicy } from "@azure/keyvault-certificates";
 import { DefaultAzureCredential } from "@azure/identity";
+
+// Load the .env file if it exists
+import * as dotenv from "dotenv";
+dotenv.config({ path: "../.env" });
 
 // This sample creates a self-signed certificate, reads it in various ways,
 // updates the tags of the certificate and finally deletes the certificate.
 
-async function main(): Promise<void> {
+export async function main(): Promise<void> {
   // If you're using MSI, DefaultAzureCredential should "just work".
   // Otherwise, DefaultAzureCredential expects the following three environment variables:
   // - AZURE_TENANT_ID: The tenant ID in Azure Active Directory
@@ -16,10 +23,13 @@ async function main(): Promise<void> {
 
   const client = new CertificateClient(url, credential);
 
-  const certificateName = "MyCertificate";
+  const certificateName = "MyCertificateHelloWorldTS";
 
   // Creating a self-signed certificate
-  const createPoller = await client.beginCreateCertificate(certificateName, DefaultCertificatePolicy);
+  const createPoller = await client.beginCreateCertificate(
+    certificateName,
+    DefaultCertificatePolicy
+  );
 
   const pendingCertificate = createPoller.getResult();
   console.log("Certificate: ", pendingCertificate);