+++ sdk/keyvault/keyvault-secrets/samples/javascript/backupAndRestore.js
@@ -1,6 +1,13 @@
-const { SecretClient } = require("../../src");
-const { DefaultAzureCredential } = require("@azure/identity");
+// Copyright (c) Microsoft corporation.
+// Licensed under the MIT license.
+
 const fs = require("fs");
+
+const { SecretClient } = require("@azure/keyvault-secrets");
+const { DefaultAzureCredential } = require("@azure/identity");
+
+// Load the .env file if it exists
+require("dotenv").config();
 
 function writeFile(filename, text) {
   return new Promise((resolve, reject) => {
@@ -35,7 +42,7 @@ async function main() {
   const url = `https://${vaultName}.vault.azure.net`;
   const client = new SecretClient(url, credential);
 
-  const secretName = "StorageAccountPassword19312312";
+  const secretName = "secretBackupAndRestoreJS";
 
   // Create our secret
   await client.setSecret(secretName, "XYZ789");