+++ sdk/textanalytics/cognitiveservices-textanalytics/rollup.base.config.js
@@ -1,6 +1,6 @@
 import path from "path";
 import nodeResolve from "rollup-plugin-node-resolve";
-import multiEntry from "rollup-plugin-multi-entry";
+import multiEntry from "@rollup/plugin-multi-entry";
 import cjs from "rollup-plugin-commonjs";
 import replace from "@rollup/plugin-replace";
 import { terser } from "rollup-plugin-terser";